Fixed up some code corruption... oO.

+++ trunk/Demoder.Common/Math.cs
@@ -8,7 +8,7 @@ namespace Demoder.Common
 	public static class Math
 	{
 		public static int Percent (int max, int cur) {
-			return (int)System.Math.Round((decimal)(cur * 100 / max), 0);
+			return (int)System.Math.Round((double)(cur * 100 / max), 0);
 		}
 	}
 }
